LayoutTests/platform/qt/http/tests/security/cookies/third-party-cookie-blocking-xslt-expected.txt

CONSOLE MESSAGE: line 2: <html xmlns='http://www.w3.org/1999/xhtml/'><body><p>Running an XSL-T 1.0 stylesheet with a 2.0 processor.</p></body></html>


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
Cookie is NOT set
